Add default VirtualHost "localhost"
This solves a issue where prosody fails to start/reload it's
configuration when there is no vhost. Adding localhost is a safe
solution according to the default config.

Ideally, interactions with the service would be queued till after
the lwrp has ran. I'll have to look into this more at a later date.